+++ test/Sentry.Tests/SentryOptionsExtensionsTests.cs
@@ -4,6 +4,7 @@ using NSubstitute;
 using Sentry.Extensibility;
 using Sentry.Integrations;
 using Sentry.Internal;
+using Sentry.Testing;
 #if NETFX
 using Sentry.PlatformAbstractions;
 #endif
@@ -271,6 +272,22 @@ namespace Sentry.Tests
         public void Integrations_Includes_TaskUnobservedTaskExceptionIntegration()
         {
             Assert.Contains(Sut.Integrations!, i => i.GetType() == typeof(TaskUnobservedTaskExceptionIntegration));
+        }
+
+        [Fact]
+        public async void SetupStartupTime_StartupTimeSet()
+        {
+            //Arrange
+            var unsetDateTime = new DateTime(1995, 01, 01);
+            var options = new SentryOptions();
+            options.StartupTime = unsetDateTime;
+            var func = new Func<bool> (()=> options.StartupTime != unsetDateTime);
+
+            //Act
+            options.SetupStartupTime();
+
+            //Assert
+            Assert.True(await func.WaitConditionAsync(true, TimeSpan.FromSeconds(1)));
         }
 
         [Theory]